No. of days fits months | refactors
