+++ src/components/NavigationBar.tsx
@@ -3,7 +3,7 @@ import NavigationButton from "./NavigationButton";
 
 function NavigationBar() {
     
-    const pages = [{ name: "H", link: "/"}, { name: "T", link: "/timestamp" }]
+    const pages = [{ name: "Home", link: "/"}, { name: "Timestamp", link: "/timestamp" }]
 
     return(
         <div className="h-fit w-full flex-shrink-0 p-2 flex flex-row bg-red-500 items-center self-center">
