+++ src/pages/timestamp/components/DateTimeField.tsx
@@ -1,7 +1,8 @@
-import { Dispatch, SetStateAction } from "react";
+import { Dispatch, SetStateAction, useEffect, useState } from "react";
 import ScrollSelector from "../../../components/scrollSelector/ScrollSelector";
-import { range } from "../../../libraries/General";
+import { getNumberOfMonth, range } from "../../../libraries/General";
 import { months } from "../../../data/Date";
+import { invoke } from "@tauri-apps/api/core";
 
 function DateTimeField({
     hour, setHour,
@@ -16,6 +17,20 @@ function DateTimeField({
     month: string, setMonth: Dispatch<SetStateAction<string>>,
     year: number, setYear: Dispatch<SetStateAction<number>>
 }) {
+
+    const [daysOfMonth, setDaysOfMonth] = useState<number>(31);
+
+    // Recalculate amount of days the month has
+    useEffect(() => {
+        invoke<number>('get_days_count_of_month', { month: getNumberOfMonth(month), year: year })
+            .then((response) => {
+                setDaysOfMonth(response);
+                if(day > response) {
+                    setDay(response);
+                }
+            });
+    }, [month]);
+
     return(
         <div className="w-full h-fit min-h-[200px] p-5 flex flex-row rounded-xl bg-bgSecondary">
             <h2 className="text-textPrimary">Time/Date</h2>
@@ -24,7 +39,7 @@ function DateTimeField({
             <p className="mt-auto mb-auto mr-1 ml-1 text-3xl text-textPrimary">:</p>
             <ScrollSelector possibleValues={range(0, 59)} selectedValue={minute} setSelectedValue={setMinute}/>
 
-            <ScrollSelector className="ml-auto" possibleValues={range(1, 31)} selectedValue={day} setSelectedValue={setDay}/>
+            <ScrollSelector className="ml-auto" possibleValues={range(1, daysOfMonth)} selectedValue={day} setSelectedValue={setDay}/>
             <ScrollSelector className="ml-2" possibleValues={months} selectedValue={month} setSelectedValue={setMonth}/>
             <ScrollSelector className="ml-2" possibleValues={range(2000, 2100)} selectedValue={year} setSelectedValue={setYear}/>
             {/*TODO - Make min and max year a setting */}
