+++ src/pages/timestamp/components/ModeField.tsx
@@ -1,10 +1,10 @@
 import { Dispatch, SetStateAction } from "react";
-import { getAllModeNames } from "../../../data/Modes";
 import SelectOneMenu from "../../../components/selectOneMenu/SelectOneMenu";
+import { getAllModeNames } from "../../../libraries/General";
 
 function ModeField({selectedMode, setSelectedMode}: {selectedMode: string, setSelectedMode: Dispatch<SetStateAction<string>>}) {
     return(
-        <div className="w-full h-fit min-h-[200px] p-5 mt-5 flex flex-row rounded-xl bg-bgSecondary">
+        <div className="w-full h-fit p-5 mt-5 flex flex-row rounded-xl bg-bgSecondary justify-center">
             <SelectOneMenu allOptions={getAllModeNames()} selectedOption={selectedMode} setSelectedOption={setSelectedMode}/>
         </div>
     )
